derp

+++ labels/kmeans2.py
@@ -34,7 +34,7 @@ def func(x):
         config.data_path, h5fn[-4], df)
     if os.path.isfile(fname):
         return True
-    pred = model.predict(image).astype(np.uint8).reshape(1024, 1024)
+    pred = model.predict(image).reshape(1024, 1024)
     imwrite(fname, pred)
     fname = "%s/kmeans_labels/%s/%s.png" % (
         config.data_path, h5fn[-4], df)
